Ensure tempDir exist before deleting it.

+++ src/main/java/net/javasauce/ss/SnowShovel.java
@@ -258,7 +258,9 @@ public class SnowShovel implements AutoCloseable {
         decompiler = new ToolProvider(this, MavenNotation.parse("net.javasauce:Decompiler:0:testframework@zip"))
                 .enableExtraction();
 
-        Files.walkFileTree(tempDir, new DeleteHierarchyVisitor());
+        if (Files.exists(tempDir)) {
+            Files.walkFileTree(tempDir, new DeleteHierarchyVisitor());
+        }
 
         // Nuke the repo if we are told to start from scratch.
         if (shouldClean && Files.exists(repoDir)) {
